4_Deployment using Fastapi/requirements.txt
fastapi
uvicorn
mlflow
scikit-learn
joblib
python-dotenv
boto3
pandas